+++ kernel/cache/cacheman.c
@@ -19,7 +19,7 @@
  */
 struct cman_node
 {
-	uint sz; /* Size of this node */
+	size_t sz; /* Size of this node */
 	struct cman_node* next; /* Next node in the list */
 };
 
@@ -28,13 +28,13 @@ static struct cman_node* head;
 void cman_init(void)
 {
 	/* Zero our space */
-	uint sz = PGROUNDUP(KVM_DISK_E - KVM_DISK_S);
+	size_t sz = PGROUNDUP(KVM_DISK_E - KVM_DISK_S);
 	memset((void*)KVM_DISK_S, 0, sz);
 	head = (void*)KVM_DISK_S;
 	head->sz = sz;
 }
 
-void* cman_alloc(uint sz)
+void* cman_alloc(size_t sz)
 {
 	if(sz == 0) return NULL;
 
@@ -87,4 +87,4 @@ void* cman_alloc(uint sz)
 	return (void*)result;
 }
 
-void cman_free(void* ptr, uint sz){}
+void cman_free(void* ptr, size_t sz){}
